reorg where fit funcs are & associated
