+++ specparam/objs/algorithm.py
@@ -94,30 +94,8 @@ class SpectralFitAlgorithm():
         self._reset_data_results(True, True, True)
 
 
-    def fit(self, freqs=None, power_spectrum=None, freq_range=None):
-        """Fit the full power spectrum as a combination of periodic and aperiodic components.
-
-        Parameters
-        ----------
-        freqs : 1d array, optional
-            Frequency values for the power spectrum, in linear space.
-        power_spectrum : 1d array, optional
-            Power values, which must be input in linear space.
-        freq_range : list of [float, float], optional
-            Frequency range to restrict power spectrum to.
-            If not provided, keeps the entire range.
-
-        Raises
-        ------
-        NoDataError
-            If no data is available to fit.
-        FitError
-            If model fitting fails to fit. Only raised in debug mode.
-
-        Notes
-        -----
-        Data is optional, if data has already been added to the object.
-        """
+    def _fit(self, freqs=None, power_spectrum=None, freq_range=None):
+        """Define the full fitting algorithm."""
 
         # If freqs & power_spectrum provided together, add data to object.
         if freqs is not None and power_spectrum is not None:
